backend/requirements.txt
fastapi
uvicorn
selenium
pytesseract
debugpy
python-dotenv
pillow
mysql-connector-python==9.3.0
pytest
pytest-asyncio
httpx
coverage
pydantic
pymysql
alembic
cryptography